+++ src/IntegrationTests/IndexFixtureBehavior.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using MyLab.Elastic.Test;
+using MyLab.Search.EsTest;
 using Xunit;
 using Xunit.Abstractions;